Add NumPy construction helpers (Wave 2) (#45)
Add the standard NumPy array-construction helpers, reusing the existing
runtime-dtype Array machinery (no new infra):

- arange(start, stop, step, dtype): the 2/3-arg form alongside the existing
  1-arg arange(n). Element count = max(0, ceil((stop-start)/step)); default
  dtype follows numpy (int64 if start/stop/step are all integers, else
  float64). Backed by a parallel affine ramp arange_range writing
  start + i*step on the NUMA workers.
- linspace(start, stop, num, endpoint, dtype): num evenly spaced points,
  float64 by default; the last point is forced exactly to stop when endpoint.
- eye(N, M, k, dtype) / identity(n, dtype): 2-D array with ones on the k-th
  diagonal; parallel for_loop over N*M cells.
- empty(shape, dtype): allocate without a value-fill (numpy.empty).
- zeros_like / ones_like / empty_like / full_like: pure-Python wrappers that
  reuse the existing constructors with a's shape and dtype.

Tests: tests/test_numpy_wave2.py (38 cases) checks values and dtype against a
numpy oracle for every helper. arange(n) now follows numpy's int64 default, so
a handful of existing tests that relied on the old float64 default request
dtype="float64" explicitly (the float arithmetic they exercise is unchanged).

check.sh green: ruff clean, 783 passed / 8 skipped, 100% coverage.

+++ tests/test_slice.py
@@ -9,7 +9,7 @@ import hpxpy as hpx
 # --- element access ---------------------------------------------------------
 
 def test_getitem_value_and_negative():
-    a = hpx.arange(10)
+    a = hpx.arange(10, dtype="float64")
     assert a[0] == 0.0
     assert a[5] == 5.0
     assert a[-1] == 9.0
